Moved some tests to specific test file
